+++ Source/WebCore/css/parser/CSSPropertyParserConsumer+NoneDefinitions.h
@@ -32,6 +32,7 @@
 
 namespace WebCore {
 
+class CSSCalcSymbolsAllowed;
 class CSSParserTokenRange;
 
 namespace CSSPropertyParserHelpers {
